extracted hook, clean up unused prop
