+++ x-pack/plugins/fleet/public/applications/fleet/sections/agents/agent_list_page/components/search_and_filter_bar.tsx
@@ -103,10 +103,7 @@ export const SearchAndFilterBar: React.FunctionComponent<{
     <>
       {isEnrollmentFlyoutOpen ? (
         <EuiPortal>
-          <AgentEnrollmentFlyout
-            agentPolicies={agentPolicies}
-            onClose={() => setIsEnrollmentFlyoutOpen(false)}
-          />
+          <AgentEnrollmentFlyout onClose={() => setIsEnrollmentFlyoutOpen(false)} />
         </EuiPortal>
       ) : null}
