+++ x-pack/plugins/fleet/public/hooks/use_agent_enrollment_flyout_data.ts
@@ -1,0 +1,38 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { useMemo } from 'react';
+
+import type { AgentPolicy } from '../types';
+
+import { useGetAgentPolicies } from '.';
+
+interface AgentEnrollmentFlyoutData {
+  agentPolicies: AgentPolicy[];
+  refreshAgentPolicies: () => void;
+}
+
+export function useAgentEnrollmentFlyoutData(): AgentEnrollmentFlyoutData {
+  const {
+    data: agentPoliciesData,
+    isLoading: isLoadingAgentPolicies,
+    resendRequest: refreshAgentPolicies,
+  } = useGetAgentPolicies({
+    page: 1,
+    perPage: 1000,
+    full: true,
+  });
+
+  const agentPolicies = useMemo(() => {
+    if (!isLoadingAgentPolicies) {
+      return agentPoliciesData?.items ?? [];
+    }
+    return [];
+  }, [isLoadingAgentPolicies, agentPoliciesData?.items]);
+
+  return { agentPolicies, refreshAgentPolicies };
+}
